cmake: use working directory for espfs.paths
cmake supports building in a directory that is different from the root
of the source repository. This is done by passing the source directory
as an argument to `cmake`.

For example:

    mkdir work
    cd work
    cmake ..
    make

Paths for commands need to be adapted to this relative directory.

The linker was already set up with this in mind, however the Python
program `pathlist.py` was writing the contents of `espfs.paths` to the
source directory, resulting in a dependency error when this was not the
same as the source directory.

This patch corrects `pathlist.py` so that it writes to the cmake working
directory rather than the source directory.

Signed-off-by: Sean Cross <sean@xobs.io>