Python package hell:
Use unreleased hiyapyco "0.4.16" from github (pip install HiYaPyCo@git+https://github.com/zerwes/hiyapyco)
to be able to use jinja2 >=3.

We need jinja2 in a version >= 3 to avoid the error with markupsafe
from markupsafe import soft_unicode".

We are forced to use jinja2 in a version <3.1 because of breaking changes
that cause problems in IDF (pip install "jinja2<3.1")

Jinja2 3.0.3 again has a problem with a breaking change in MarkupSafe 2.1.x.
Therefore limit the version to <2.1.0 in this requirements file.